swarops2-main/backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
annotated-doc
annotated-types
anyio
appdirs
attrs
backoff
bcrypt
beautifulsoup4
black
boto3
botocore
build
certifi
cffi
charset-normalizer
chromadb
click
crewai
crewai-tools
cryptography
defusedxml
deprecation
diskcache
distro
dnspython
docker
docstring_parser
durationpy
ecdsa
email-validator

et_xmlfile
fastapi
fastuuid
filelock
flake8
flatbuffers
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface_hub
idna
importlib_metadata
importlib_resources
iniconfig
instructor
isort
Jinja2
jiter
jmespath
jq
json5
json_repair
jsonpatch
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
lance-namespace
lance-namespace-urllib3-client
lancedb
langchain
langchain-core
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
librt
linkify-it-py
litellm
lxml
markdown-it-py
MarkupSafe
mccabe
mcp
mdit-py-plugins
mdurl
mmh3
motor
mpmath
multidict
mypy
mypy_extensions
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
passlib
pathspec
pdfminer.six
pdfplumber
pillow
platformdirs
pluggy
portalocker
posthog
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pybase64
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
PyMuPDF
pyparsing
pypdfium2
PyPika
pyproject_hooks
pytest
python-dateutil
python-docx
python-dotenv
python-jose
python-multipart
pytokens
pytube
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
resend
rich
rpds-py
rsa
s3transfer
s5cmd
shellingham
six
sniffio
soupsieve
sse-starlette
starlette
stripe
sympy
tenacity
textual
tiktoken
tokenizers
tomli
tomli_w
tqdm
typer
typing-inspection
typing_extensions
tzdata
uc-micro-py
uritemplate
urllib3
uuid_utils
uv
uvicorn

watchfiles
websocket-client
websockets
xxhash
yarl
youtube-transcript-api
zipp
zstandard
